+++ src/design-system/index.ts
@@ -8,6 +8,9 @@
 
 // Atoms (primitives-layout/)
 export { Container, type ContainerProps } from './primitives-layout/container'
+export { Stack, type StackProps } from './primitives-layout/stack'
+export { Cluster, type ClusterProps } from './primitives-layout/cluster'
+export { Grid, type GridProps } from './primitives-layout/grid'
 
 // Atoms (core/)
 export { Button, type ButtonProps } from './core/button'